NCBIm38 library: replace BLAST binary DBs, correct some coordinates
The custom mouse database (based on the NCBIm38 assembly) previously
shipped as FASTA files and pre-compiled binary BLAST databases. This
makes changes to the database intransparent since a lot of files are
affected. Exchange these files with CSVs containing the coordinates
and a download script. Adapt the respective "build_" script
accordingly.

It was further noted that the previously used coordinates of Igk-V
family 4 were 6 bp short on the 5' side. Provide corrected tables.