Correct auxiliary data file for NCBIm38 database
IgBLAST uses an auxiliary data file (optional_file/*.aux) to determine
the correct reading frame of J segment to assess the productivity
of the V(D)J joint. The identifiers in this file have to match
exactly the identifiers in the BLAST DB.

Provide a corrected version of the .aux file. Adapt the respective
"build_" script to handle the changes correctly.